GCF file split: concepts
