+++ functions/concepts/env-vars/src/main/java/functions/EnvVars.java
@@ -1,0 +1,41 @@
+/*
+ * Copyright 2020 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package functions;
+
+// [START functions_env_vars]
+
+import com.google.cloud.functions.HttpFunction;
+import com.google.cloud.functions.HttpRequest;
+import com.google.cloud.functions.HttpResponse;
+import java.io.BufferedWriter;
+import java.io.IOException;
+
+public class EnvVars implements HttpFunction {
+
+  // Returns the environment variable "foo" set during function deployment.
+  @Override
+  public void service(HttpRequest request, HttpResponse response)
+      throws IOException {
+    BufferedWriter writer = response.getWriter();
+    String foo = System.getenv("FOO");
+    if (foo == null) {
+      foo = "Specified environment variable is not set.";
+    }
+    writer.write(foo);
+  }
+}
+// [END functions_env_vars]
